% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_classify.r
\docType{methods}
\name{ground_classify}
\alias{ground_classify}
\alias{ground_classify,LAS-method}
\title{Classify points as ground or not ground}
\usage{
ground_classify(obj, MaxWinSize = 20, slope = 1, InitDist = 0.5,
  MaxDist = 3)

\S4method{ground_classify}{LAS}(obj, MaxWinSize = 20, slope = 1,
  InitDist = 0.5, MaxDist = 3)
}
\arguments{
\item{obj}{LAS objet}

\item{MaxWinSize}{param of the algorithm}

\item{slope}{param of the algorithm}

\item{InitDist}{param of the algorithm}

\item{MaxDist}{param of the algorithm}
}
\value{
Nothing. The original object is updated by reference. The 'Classification'
column contains 1 for non-ground and 2 for ground according to LAS specifications.
}
\description{
This function is a wrapper for the Progressive Morphological Filter for segmentation
of ground points implemented in PCL (Point Cloud Library, see references).
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")

lidar = readLAS(LASfile)

ground_classify(lidar)
plot(lidar, color = "Classification")
}
}
\references{
Zhang, K., Chen, S. C., Whitman, D., Shyu, M. L., Yan, J., & Zhang, C. (2003). A progressive morphological filter for removing nonground measurements from airborne LIDAR data. IEEE Transactions on Geoscience and Remote Sensing, 41(4 PART I), 872–882. http://doi.org/10.1109/TGRS.2003.810682
R. B. Rusu and S. Cousins, "3D is here: Point Cloud Library (PCL)," Robotics and Automation (ICRA), 2011 IEEE International Conference on, Shanghai, 2011, pp. 1-4. doi: 10.1109/ICRA.2011.5980567
}
\seealso{
\link[lidR:LAS-class]{LAS}
}

