% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcl_ground_classify}
\alias{pcl_ground_classify}
\title{Internal version of ground_classify}
\usage{
pcl_ground_classify(X, Y, Z, MaxWinSize = 20L, slope = 1, InitDist = 0.5,
  MaxDist = 3)
}
\description{
Internal version of ground_classify
}

